Added a fix for conversion of mods with options

+++ bcml/dev.py
@@ -574,9 +574,12 @@ def _convert_sarc(sarc: oead.Sarc, to_wiiu: bool) -> Tuple[bytes, List[str]]:
 def convert_mod(mod: Path, to_wiiu: bool, warn_only: bool = False) -> list:
     warnings = []
 
-    def handle_warning(warning: str) -> None:
+    def handle_warning(warning: Union[str, list]) -> None:
         if not warn_only:
             raise ValueError(warning)
+        elif isinstance(warning, list):
+            for warning_ in warning:
+                warnings.append(warning_)
         else:
             warnings.append(warning)
 
@@ -597,14 +600,14 @@ def convert_mod(mod: Path, to_wiiu: bool, warn_only: bool = False) -> list:
 
     special_files = {"ActorInfo.product.sbyml"}
     all_files = {
-        f for f in mod.rglob("**/*.*") if f.is_file() and "options" not in f.parts
+        f for f in mod.rglob("*.*") if f.is_file() and f not in (mod / "options").rglob("*.*")
     }
 
     for file in all_files:
         if file.suffix in NO_CONVERT_EXTS:
             handle_warning(
                 "This mod contains a file which the platform converter does not support:"
-                f" {str(file.relative_to(mod))}"
+                f" {str(file.relative_to(mod)) if 'options' not in mod.parts else str(file.relative_to(mod.parent))}"
             )
 
     actorinfo_log = mod / "logs" / "actorinfo.yml"
@@ -797,9 +800,12 @@ def convert_mod(mod: Path, to_wiiu: bool, warn_only: bool = False) -> list:
 
     if (mod / "options").exists():
         for folder in {d for d in (mod / "options").glob("*") if d.is_dir()}:
-            convert_mod(folder, to_wiiu=to_wiiu, warn_only=warn_only)
+            option_warnings = convert_mod(folder, to_wiiu=to_wiiu, warn_only=warn_only)
+            handle_warning(option_warnings)
     
-    meta = loads((mod / "info.json").read_text("utf-8"))
-    meta["platform"] = "wiiu" if to_wiiu else "switch"
-    (mod / "info.json").write_text(dumps(meta, indent=2, ensure_ascii=False), "utf-8")
+    if (mod / "info.json").exists():
+        meta = loads((mod / "info.json").read_text("utf-8"))
+        meta["platform"] = "wiiu" if to_wiiu else "switch"
+        (mod / "info.json").write_text(dumps(meta, indent=2, ensure_ascii=False), "utf-8")
+
     return warnings
